Implement suggestions from @ogoffart
